#67 Вынос логики в отдельные файлы
